% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cluster6}
\alias{cluster6}
\title{Word Frequencies of 6 News Groups Clusters}
\format{A list with two elements: \describe{ \item{data}{A data frame with
  413588 rows and 3 variables. The \code{row} and \code{column} variables are
  the respective indices of the nonzero entries of the sparse matrix. Rows
  correspond to documents from the corpus; columns correspond to a word from
  the vocabulary. The \code{value} column is the frequency of the word in
  that document.} \item{labels}{A integer vector with 3189 values ranging
  from 1 to 6 corresponding to the ground truth labels of the cluster}}}
\source{
\url{http://qwone.com/~jason/20Newsgroups/}
}
\usage{
cluster6
}
\description{
A dataset containing the word frequencies and cluster labels of 6 news groups
selected from the entire 20 news groups dataset. The 6 groups were selected
by taking the first listed topic within each of the six larger subject areas.
The selected groups are (number corresponds to cluster label): enumerate{
\item comp.graphics \item misc.forsale \item rec.autos \item sci.crypt \item
talk.politics.misc \item talk.religion.misc }
}
\keyword{datasets}
