% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cluster6_counts}
\alias{cluster6_counts}
\title{Word Frequencies of 6 News Groups Clusters}
\format{A data frame with 413588 rows and 3 variables: \describe{
  \item{row}{row index; corresponds to a document from the corpus}
  \item{col}{column index; corresponds to a word from the vocabulary}
  \item{value}{frequency of the word in that document} }}
\source{
\url{http://qwone.com/~jason/20Newsgroups/}
}
\usage{
cluster6_counts
}
\description{
A dataset containing the word frequencies and cluster labels of 6 news groups
selected from the entire 20 news groups dataset. The 6 groups were selected by
taking the first listed topic within each of the six larger subject areas.
The selected groups are (number corresponds to cluster label):
enumerate{
  \item comp.graphics
  \item misc.forsale
  \item rec.autos
  \item sci.crypt
  \item talk.politics.misc
  \item talk.religion.misc
}
}
\keyword{datasets}
