% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cluster6_labels}
\alias{cluster6_labels}
\title{Labels of 6 News Groups Clusters}
\format{A integer vector with 3189 values ranging from 1 to 6.}
\source{
\url{http://qwone.com/~jason/20Newsgroups/}
}
\usage{
cluster6_labels
}
\description{
A dataset containing the cluster labels of 6 news groups selected from the
entire 20 news groups dataset. The 6 groups were selected by taking the first
listed topic within each of the six larger subject areas. The selected groups
are (number corresponds to cluster label):
enumerate{
  \item comp.graphics
  \item misc.forsale
  \item rec.autos
  \item sci.crypt
  \item talk.politics.misc
  \item talk.religion.misc
}
}
\keyword{datasets}
