% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustercheck.R
\name{clustercheck}
\alias{clustercheck}
\title{Checking cluster accuracy}
\usage{
clustercheck(cluster, Labels, k)
}
\arguments{
\item{k}{}
}
\description{
in the results table, true clusters are the rows, guessed clusters are the columns
}
