% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering}
\alias{clustering}
\title{Spectral clustering}
\usage{
clustering(Weights, method, k, t = NULL, sparse = T)
}
\arguments{
\item{method}{NJW, Ncut, DiffusionMap}

\item{sparse}{}
}
\description{
Spectral clustering
}
