% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colweights.R
\name{colweights}
\alias{colweights}
\title{Column weighting}
\usage{
colweights(data, weightfunction, sparseinput, par1 = NULL, par2 = NULL,
  mode = NULL, binary = TRUE, convertsparse = TRUE)
}
\arguments{
\item{data}{Matrix of counts.}

\item{weightfunction}{One of the following: beta", "step", "linear", "IDF", "IDF^2".}

\item{sparseinput}{Logical. Is the input data sparse?}

\item{par1}{Beta parameters or step function boundaries.}

\item{par2}{Beta parameters or step function boundaries.}

\item{mode}{Linear function max (desired density max weight, e.g. 1/k).}

\item{binary}{Logical. Convert data to binary?}

\item{convertsparse}{Logical. If your matrix is dense, convert to sparse}
}
\value{
Processed version of original input matrix.
}
\description{
Column weighting
}
