% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInsights.R
\name{getInsights}
\alias{getInsights}
\title{Determine the Top-N Keywords Per Cluster}
\usage{
getInsights(cluster, vocab, n, plot = FALSE, file = NULL)
}
\arguments{
\item{cluster}{Data from one of the clusters.}

\item{vocab}{Column names, like vocabulary, websites. Vocab should be one to
one with column indices.}

\item{n}{Top n features you would like to see.}

\item{plot}{Logical. To produce plot in addition to outputting top-n keywords and
maginitudes.}

\item{file}{Filepath and PNG file name. Note: plot must be set to TRUE to save a
file.}
}
\value{
Barchart of the top-n influential keywords in the cluster.
}
\description{
Determine the Top-N Keywords Per Cluster
}
