% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights.R
\name{insights}
\alias{insights}
\title{Determine the Top-N Keywords Per Cluster}
\usage{
insights(data, labels, cluster, top_n = 10, plot = TRUE, file = NULL)
}
\arguments{
\item{labels}{Integer vector. Labels output by `clustering()`}

\item{cluster}{Integer. Cluster label for which to calculate top keywords.}

\item{top_n}{Integer. Specificies the number of top keywords to show.}

\item{plot}{Logical. If TRUE, a barplot of the keyword importance is returned.}
}
\value{
List of 2. First element: keywords in decreasing order of importance
  Second element: maginitude of importance associated with the keyword.
}
\description{
Determine the Top-N Keywords Per Cluster
}
