% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunction.R
\name{mainfunction}
\alias{mainfunction}
\title{Main Function to Run Complete Algorithm}
\usage{
mainfunction(data, nclust, sparse = TRUE, convertsparse = TRUE,
  save = TRUE, return = TRUE, preset = 0, weightfunction = "IDF",
  binary = TRUE, lower = 2, upper = NULL, par1 = NULL, par2 = NULL,
  mode = NULL, simfunction = "cosine", simscale = NULL,
  rowscaling = NULL, colscaling = NULL, sigma = NULL, centers = NULL,
  seed = NULL, distance = NULL, clusterfunction = "DiffusionMap",
  t = 0.5, kmeans.method = "kmeans", m = NULL, weight.SVD = FALSE,
  SVDdim = 200, SVDprint = FALSE, dim1 = 1, dim2 = 2, dim3 = 3,
  filepath = NULL, SVDsim = TRUE, simdim = 3, dim1sim = 1,
  dim2sim = 2, dim3sim = 3, SVDsim.plot = FALSE, sim.filepath = NULL,
  insights = FALSE, vocab = NULL, nfeatures = 20, n.ins = NULL,
  insight.plot = TRUE, insight.filepath = NULL, end.of.arguments = NULL)
}
\arguments{
\item{data}{See description.}

\item{nclust}{See description.}

\item{sparse}{See description.}

\item{convertsparse}{See description.}

\item{save}{See description.}

\item{return}{See description.}

\item{preset}{See description.}

\item{weightfunction}{See description.}

\item{binary}{See description.}

\item{lower}{See description.}

\item{upper}{See description.}

\item{par1}{See description.}

\item{par2}{See description.}

\item{mode}{See description.}

\item{simfunction}{See description.}

\item{simscale}{See description.}

\item{rowscaling}{See description.}

\item{colscaling}{See description.}

\item{sigma}{See description.}

\item{centers}{See description.}

\item{seed}{See description.}

\item{distance}{See description.}

\item{clusterfunction}{See description.}

\item{t}{See description.}

\item{kmeans.method}{See description.}

\item{m}{See description.}

\item{weight.SVD}{See description.}

\item{SVDdim}{See description.}

\item{SVDprint}{See description.}

\item{dim1}{See description.}

\item{dim2}{See description.}

\item{dim3}{See description.}

\item{filepath}{See description.}

\item{SVDsim}{See description.}

\item{simdim}{See description.}

\item{dim1sim}{See description.}

\item{dim2sim}{See description.}

\item{dim3sim}{See description.}

\item{SVDsim.plot}{See description.}

\item{sim.filepath}{See description.}

\item{insights}{See description.}

\item{vocab}{See description.}

\item{nfeatures}{See description.}

\item{n.ins}{See description.}

\item{insight.plot}{See description.}

\item{insight.filepath}{See description.}

\item{end.of.arguments}{See description.}
}
\value{
List.
}
\description{
The accompanying pdf has all the details for this function. Please see this
document for details about all function arguments.
}
