% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtsne_viz.R
\name{rtsne_viz}
\alias{rtsne_viz}
\title{Rtsne visualization}
\usage{
rtsne_viz(data, labels)
}
\arguments{
\item{data}{Original data.}

\item{labels}{Integer vector. Can be true labels or labels created by clustering output.}
}
\value{
plot
}
\description{
Rtsne visualization
}
