% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity}
\alias{similarity}
\title{Calculate similarity matrices}
\usage{
similarity(data, method, rowscaling = NULL, colscaling = NULL,
  sigma = NULL, centers = NULL, seed = NULL, distance = NULL,
  sparse = T)
}
\arguments{
\item{data}{matrix}

\item{method}{"correlation" "cosine" "dotproduct" "Gaussian"}

\item{rowscaling}{"L2" "L1"}

\item{colscaling}{"standardize" (center and divide by variance)}

\item{sigma}{variance for Gaussian Kernel}

\item{centers}{% of columns to be sampled, if desired (e.g. centers = .25)}

\item{seed}{for repeatability}

\item{distance}{for use with Gaussian Kernel; options = JSdivergence, L1, L2}

\item{sparse}{logical. Is the data sparse or not.}
}
\description{
Calculate similarity matrices
}
