% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdProcessing.R
\name{svdProcessing}
\alias{svdProcessing}
\title{Dimensionality Reduction Via SVD and 3-D Plotting}
\usage{
svdProcessing(data, n = 10, scale = 1, plot = FALSE, file = NULL)
}
\arguments{
\item{data}{Matrix object. Accepts sparse or dense form.}

\item{n}{Number of dimensions to be reduced to (default is 10)}

\item{scale}{The sclae of axis, in case the data is very sparse. If data is
too sparse, try 1000, ..., 10^6, etc.}

\item{plot}{Logical. Return 3-D plots}

\item{file}{Full path to filename without file type, which will be PNG.}
}
\value{
Matrix object with the same number of rows as the original input
  data, and a reduced number of columns equal to the input argument \code{n}.
}
\description{
Dimensionality Reduction Via SVD and 3-D Plotting
}
