% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colweights.R
\name{colweights}
\alias{colweights}
\title{Column weighting}
\usage{
colweights(data, weightfunction, par1 = NULL, par2 = NULL, mode = NULL,
  binary = TRUE, sparse = TRUE)
}
\arguments{
\item{weightfunction}{"beta" "step" "linear" "IDF".}

\item{par1}{beta parameters or step function boundaries.}

\item{par2}{beta parameters or step function boundaries.}

\item{mode}{linear function max (desired density max weight, e.g. 1/k).}

\item{binary}{logical. Is data binary or not.}

\item{sparse}{logical. Is data sparse or not.}
}
\description{
Column weighting
}
