% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colweights.R
\name{colweights}
\alias{colweights}
\title{Column weighting}
\usage{
colweights(data, weightfunction, sparseinput, par1 = NULL, par2 = NULL,
  mode = NULL, binary = TRUE, convertsparse = TRUE, lower = 2,
  upper = NULL)
}
\arguments{
\item{data}{Matrix of counts.}

\item{weightfunction}{One of the following: beta", "step", "linear", "IDF",
"IDF^2".}

\item{sparseinput}{Logical. Is your data in sparse (long) format? (Long
format = Nx3 matrix-like object of nonzero entries)}

\item{par1}{Beta parameters or step function boundaries.}

\item{par2}{Beta parameters or step function boundaries.}

\item{mode}{Linear function max (desired density max weight, e.g. 1/k).}

\item{binary}{Logical. Convert data to binary?}

\item{convertsparse}{Logical. If your matrix is dense, convert to sparse}

\item{lower}{Integer. Lower bound for column sum in order to retain column;
columns with sums below this threshold will be removed.}

\item{upper}{Integer. Upper bound for column sum in order to retain column;
columns with sums above this threshold will be removed.}
}
\value{
Processed version of original input matrix.
}
\description{
Column weighting
}
\seealso{
\code{\link{similarity}}
}
