% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news_subset.R
\name{news_subset}
\alias{news_subset}
\title{Select a subset of the full 20Newsgroupds dataset}
\usage{
news_subset(X, filter, binary = TRUE, vocabulary = FALSE)
}
\arguments{
\item{vocabulary}{}
}
\description{
Select a subset of the full 20Newsgroupds dataset
}
