% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity}
\alias{similarity}
\title{Calculate similarity matrices}
\usage{
similarity(data, method, rowscaling = NULL, colscaling = NULL,
  sigma = NULL, centers = NULL, seed = NULL, distance = NULL,
  sparse = TRUE)
}
\arguments{
\item{data}{Matrix.}

\item{method}{One of the following: "correlation", "cosine", "dotproduct", "Gaussian".}

\item{rowscaling}{One of the following: "L2", "L1".}

\item{colscaling}{Single option: "standardize" (center and divide by variance).}

\item{sigma}{Variance for Gaussian Kernel method.}

\item{centers}{Proportion of columns to be sampled, if desired (e.g. centers = .25).}

\item{seed}{For sampling reproducibility.}

\item{distance}{For use with Gaussian Kernel; options = "JSdivergence", "L1", "L2"}

\item{sparse}{Logical. Is the data sparse or not.}
}
\value{
Similarity matrix.
}
\description{
Note: this function sets the self-similarity diagonal = 0. For certain
applications, diagonal = 1 may be more appropriate.
}
